apps/roaming-charges/requirements.txt
fastapi==0.97.0
python-dotenv==1.0.0
uvicorn==0.25.0
psycopg2-binary
langchain
langchain-ibm
ibm_watson_machine_learning